Create SQL Queries Objective
New York Restaurants
We have put together a table of restaurants called nomnom and we need your help to answer some questions. Use the SQL commands you just learned and find the best dinner spots in the city.

1) Start by getting a feel for the nomnom table:

What are the column names?

2) What are the distinct neighborhoods?

3) What are the distinct cuisine types?

4) Suppose we would like some Chinese takeout.

What are our options?

5) Return all the restaurants with reviews of 4 and above.

6) Suppose Abbi and Ilana want to have a fancy dinner date.

Return all the restaurants that are Italian and $$$.

7) Your coworker Trey can’t remember the exact name of a restaurant he went to but he knows it contains the word ‘meatball’ in it.

Can you find it for him using a query?

8) Let’s order delivery to the house!

Find all the close by spots in Midtown, Downtown or Chinatown.

(OR can be used more than once)

9) Find all the health grade pending restaurants (empty values).

10) Create a Top 10 Restaurants Ranking based on reviews.

11) Use a CASE statement to change the rating system to:

review > 4.5 is Extraordinary
review > 4 is Excellent
review > 3 is Good
review > 2 is Fair
Everything else is Poor
Don’t forget to rename the new column!