Create SQL Aggregate functions Objective
Trends in Startups
Howdy! It’s your first day as a TechCrunch reporter. Your first task is to write an article on the rising trends in the startup world.

To get you started with your research, your boss emailed you a project.sqlite file that contains a table called startups. It is a portfolio of some of the biggest names in the industry.

Write queries with aggregate functions to retrieve some interesting insights about these companies.

Tasks

1) Getting started, take a look at the startups table:

SELECT *
FROM startups;

How many columns are there?

2) Calculate the total number of companies in the table.

3) We want to know the total value of all companies in this table.

Calculate this by getting the SUM() of the valuation column.

4) What is the highest amount raised by a startup?

Return the maximum amount of money raised.

5) Edit the query so that it returns the maximum amount of money raised, during ‘Seed’ stage.

6) In what year was the oldest company on the list founded?

7) Let's find out the valuations among different sectors:

8) Return the average valuation.

9) Return the average valuation, in each category.

10) Return the average valuation, in each category.
Round the averages to two decimal places.

11) Return the average valuation, in each category.
Round the averages to two decimal places.
Lastly, order the list from highest averages to lowest.

What are the most competitive markets?

12) First, return the name of each category with the total number of companies that belong to it. Next, filter the result to only include categories that have more than three companies in them.

What are the most competitive markets?

13) Let's see if there's a difference in startups sizes among different locations:

What is the average size of a startup in each location?

14) What is the average size of a startup in each location, with average sizes above 500?